Move out navbar to common
