+++ simplq/src/components/common/Footer.jsx
@@ -31,6 +31,8 @@ export default () => (
       <div className={styles['card']}>
         <div className={styles['card-header']}>
           <div className={styles['center-horizontally']}>
+            {' '}
+            {/* TOFIX: No such class / wrong approach */}
             <h1>Open source</h1>
           </div>
         </div>
@@ -52,6 +54,8 @@ export default () => (
       <div className={styles['card']}>
         <div className={styles['card-header']}>
           <div className={styles['center-horizontally']}>
+            {' '}
+            {/* TOFIX: No such class / wrong approach */}
             <h1>Keep in touch</h1>
           </div>
         </div>
