+++ simplq/src/components/common/Nav/Burger.jsx
@@ -1,7 +1,7 @@
 /* eslint-disable jsx-a11y/click-events-have-key-events */
 /* eslint-disable jsx-a11y/no-static-element-interactions */
 import React, { useState } from 'react';
-import styles from '../../../../styles/homePage.module.scss';
+import styles from '../../../styles/navbar.module.scss';
 import LeftNav from './LeftNav';
 
 const Burger = () => {
