+++ simplq/src/components/common/Nav/Navbar.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import styles from '../../../../styles/homePage.module.scss';
+import styles from '../../../styles/navbar.module.scss';
 import Burger from './Burger';
-import Logo from '../../../common/ClickableLogo';
-import { smoothScrollTo } from '../../../common/utilFns';
+import Logo from '../ClickableLogo';
+import { smoothScrollTo } from '../utilFns';
 
 const Navbar = () => {
   const scrollToTop = () => {
