+++ simplq/src/components/pages/Home/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BenefitsInfo, HowToCreate, HowToJoin, ExtraInfo } from './StaticInfos';
 import LandingPage from './LandingPage';
-import Navbar from './Nav/Navbar';
+import Navbar from '../../common/Nav/Navbar';
 
 const Home = (props) => {
   return (
