+++ simplq/src/styles/homePage.module.scss
@@ -192,131 +192,3 @@
 #curved-div-2 {
   display: flex;
 }
-
-.burger {
-  width: 2rem;
-  height: 2rem;
-  position: fixed;
-  top: 0.5rem;
-  right: 1rem;
-  z-index: 5;
-  display: none;
-  cursor: pointer;
-  @media (max-width: 768px) {
-    display: flex;
-    justify-content: space-around;
-    flex-flow: column nowrap;
-  }
-  div {
-    width: 2rem;
-    height: 0.25rem;
-    background-color: #3a3768;
-    &[open] {
-      background-color: #fff;
-    }
-    border-radius: 0.5rem;
-    transform-origin: 1px;
-    transition: all 0.3s linear;
-
-    &:nth-child(1) {
-      transform: rotate(0);
-      &[open] {
-        transform: rotate(45deg);
-      }
-    }
-    &:nth-child(2) {
-      transform: translateX(0);
-      opacity: 1;
-      &[open] {
-        transform: translateX(100%);
-        opacity: 0;
-      }
-    }
-    &:nth-child(3) {
-      transform: rotate(0);
-      &[open] {
-        transform: rotate(-45deg);
-      }
-    }
-  }
-}
-
-.left-nav {
-  list-style: none;
-  display: flex;
-  flex-flow: row nowrap;
-  margin: 0;
-  position: fixed;
-  top: 1.4rem;
-  li {
-    padding: 0 1.5rem;
-  }
-  li:last-child {
-    @media (min-width: 787px) {
-      position: fixed;
-      top: 1.2rem;
-      right: 6rem;
-    }
-  }
-  a {
-    text-decoration: none;
-    color: grey;
-    cursor: pointer;
-  }
-  @media (max-width: 768px) {
-    flex-flow: column nowrap;
-    background-color: #3a3768;
-    position: fixed;
-    top: 0;
-    right: 0;
-    height: 100vh;
-    width: 15rem;
-    padding-top: 3.5rem;
-    transition: transform 0.3s ease-in-out;
-    left: auto;
-    opacity: 0;
-    li {
-      padding: 1.5rem 0;
-    }
-    a {
-      color: #fff;
-    }
-    transform: translateX(100%);
-    &[open] {
-      opacity: 1;
-      transform: translateX(0);
-    }
-  }
-}
-
-.left-nav-overlay {
-  display: none;
-  @media (max-width: 768px) {
-    &[open] {
-      display: block;
-      height: 100vh;
-      width: 100vw;
-      position: fixed;
-      top: 0;
-      left: 0;
-      background: rgba(0, 0, 0, 0.5);
-    }
-  }
-}
-
-.navbar {
-  width: 100%;
-  height: 4rem;
-  background-color: #fff;
-  border-bottom: 0.2rem solid #f1f1f1;
-  padding: 0 3rem;
-  z-index: 99;
-  position: fixed;
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.center-horizontally {
-  display: flex;
-  align-items: center;
-}
